arkad-sql-demo/fundamentals/pyproject.toml
[tool.poetry]
name = "fundamentals"
version = "0.1.0"
description = ""
authors = ["Oleh <oleg.davydiuk1995@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
pandas = "^2.2.0"
pandas-datareader = "^0.10.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
